+++ src/shared/lib/index.ts
@@ -1,4 +1,5 @@
 export { cn } from './cn'
+export { dateFormatting } from './dateFormatting'
 export * from './errorHandlers'
 export * from './langSwitcher'
 export { getProvidesTags } from './providesTags'
@@ -9,5 +10,6 @@ export { sleep } from './sleep'
 export { useChangeZodErrorLang } from './useChangeZodErrorLang'
 export { useInfiniteScroll } from './useInfiniteScroll'
 export * from './useProgressBar'
+export { useScreenHeightTracker } from './useScreenHeightTracker'
 export { type TranslateFuncType, useTranslate } from './useTranslate'
 export { isImgFileTypeValid } from './validateFileExtension'
