+++ src/shared/ui/index.ts
@@ -18,6 +18,8 @@ export { Dropzone, type DropzoneRef } from './dropzone/Dropzone'
 export { ImageCarousel } from './imageCarousel/ImageCarousel'
 export { Input } from './input/input'
 export { Loader } from './loader/Loader'
+export { Pagination } from './pagination/Pagination'
+export { RadioGroup, type RadioGroupProps } from './radioGroup/index'
 export { Option } from './select/option'
 export { Select } from './select/select'
 export { Slider } from './slider/slider'
