--- src/widgets/PaymentTable/index.ts
@@ -1,1 +1,0 @@
-export { PaymentTable } from './ui/PaymentTable'
