+++ src/widgets/settings-tabs/index.ts
@@ -1,0 +1,3 @@
+export { AccountManagement } from './account-management/index'
+export { MyPayments } from './my-payments/index'
+export { UserSettings } from './user-settings/index'
