+++ src/widgets/settings-tabs/my-payments/lib/formatPaymentType.ts
