+++ src/widgets/settings-tabs/my-payments/lib/formatSubscriptionType.ts
