+++ src/widgets/settings-tabs/my-payments/ui/MyPayments.module.scss
@@ -1,5 +1,3 @@
-@import '@/app/styles/mixins';
-
 .paginationSize {
   width: 50%;
 }
