+++ src/widgets/settings-tabs/my-payments/ui/skeleton/Skeleton.tsx
@@ -1,4 +1,4 @@
-import s from '@/widgets/settings-tabs/my-payments/ui/skeleton/Skeleton.module.scss'
+import s from './Skeleton.module.scss'
 
 export const Skeleton = () => {
   return <div className={s.skeleton}></div>
