+++ src/widgets/settings-tabs/user-settings/index.ts
@@ -1,0 +1,1 @@
+export { UserSettings } from './ui/UserSettings'
