+++ src/widgets/settings-tabs/user-settings/ui/UserSettings.module.scss
